# api/openapi_new.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portcullis",
    "version": "0.8.0"
  },
  "paths": {
...
    "/admin/platform/license": {
      "get": {
        "tags": [
          "admin-license"
        ],
        "summary": "Admin Get License",
        "description": "Return the current platform license for the default tenant, or None.",
        "operationId": "admin_get_license_admin_platform_license_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LicenseView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Admin Get License Admin Platform License Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-license"
        ],
        "summary": "Admin Issue License",
        "description": "Issue a new platform license (platform admin only).\n\nThe plaintext license key is returned once in the response and never\nstored; only its hash is persisted.",
        "operationId": "admin_issue_license_admin_platform_license_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LicenseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/platform/license/revoke": {
      "post": {
        "tags": [
          "admin-license"
        ],
        "summary": "Admin Revoke License",
        "description": "Revoke the current active platform license, if any (platform admin only).",
        "operationId": "admin_revoke_license_admin_platform_license_revoke_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/license": {
      "get": {
        "tags": [
          "license"
        ],
        "summary": "Org Get License",
        "description": "Return the org's current license entitlements with usage, or None (admin only).\n\nRead-only \u2014 never exposes the license key itself.  Seat usage counts the\ntenant's org-member rows; server usage counts registered servers.",
        "operationId": "org_get_license_v1_license_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LicenseUsageView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Org Get License V1 License Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tenants/{tenant_id}/license": {
      "get": {
        "tags": [
          "admin-license"
        ],
        "summary": "Tenant Get License",
        "description": "Return a managed tenant's license entitlements with usage (platform admin only).",
        "operationId": "tenant_get_license_admin_tenants__tenant_id__license_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LicenseUsageView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Tenant Get License Admin Tenants  Tenant Id  License Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-license"
        ],
        "summary": "Tenant Issue License",
        "description": "Issue a new license for a managed tenant (platform admin only).\n\nThis is the Cloud control plane's hook for seat/billing-tier changes: the\ncontrol plane sets the plan, seat_limit, and term on the tenant's license.",
        "operationId": "tenant_issue_license_admin_tenants__tenant_id__license_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LicenseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tenants/{tenant_id}/license/revoke": {
      "post": {
        "tags": [
          "admin-license"
        ],
        "summary": "Tenant Revoke License",
        "description": "Revoke a managed tenant's current active license (platform admin only).",
        "operationId": "tenant_revoke_license_admin_tenants__tenant_id__license_revoke_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LicenseCreate": {
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/LicensePlan"
          },
          "licensee": {
            "type": "string",
            "title": "Licensee"
          },
          "seat_limit": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Seat Limit"
          },
          "server_limit": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Limit"
          },
          "expires_in_days": {
            "type": "integer",
            "maximum": 3650.0,
            "exclusiveMinimum": 0.0,
            "title": "Expires In Days"
          }
        },
        "type": "object",
        "required": [
          "plan",
          "licensee",
          "seat_limit",
          "expires_in_days"
        ],
        "title": "LicenseCreate",
        "description": "Schema for issuing a platform license (platform admin only)."
      },
      "LicenseIssueResponse": {
        "properties": {
          "license": {
            "$ref": "#/components/schemas/LicenseView"
          },
          "key": {
            "type": "string",
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "license",
          "key"
        ],
        "title": "LicenseIssueResponse",
        "description": "Issue response containing the one-time plaintext license key."
      },
      "LicensePlan": {
        "type": "string",
        "enum": [
          "community",
          "pro",
          "enterprise"
        ],
        "title": "LicensePlan",
        "description": "Commercial plan tiers for the platform license."
      },
      "LicenseStatus": {
        "type": "string",
        "enum": [
          "active",
          "expired",
          "revoked"
        ],
        "title": "LicenseStatus",
        "description": "Lifecycle state of a platform license.\n\n``active``   \u2014 valid and in force (not expired, not revoked).\n``expired``  \u2014 past its expires_at timestamp.\n``revoked``  \u2014 invalidated by a platform admin."
      },
      "LicenseUsageView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "plan": {
            "$ref": "#/components/schemas/LicensePlan"
          },
          "status": {
            "$ref": "#/components/schemas/LicenseStatus"
          },
          "seat_limit": {
            "type": "integer",
            "title": "Seat Limit"
          },
          "server_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Limit"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "licensee": {
            "type": "string",
            "title": "Licensee"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "seats_used": {
            "type": "integer",
            "title": "Seats Used",
            "default": 0
          },
          "servers_used": {
            "type": "integer",
            "title": "Servers Used",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan",
          "status",
          "seat_limit",
          "server_limit",
          "issued_at",
          "expires_at",
          "revoked_at",
          "licensee",
          "created_at",
          "updated_at"
        ],
        "title": "LicenseUsageView",
        "description": "License entitlements plus live usage counts for the tenant.\n\n``seats_used`` counts the tenant's org-member rows; ``servers_used``\ncounts registered MCP servers.  Both are computed at read time."
      },
      "LicenseView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "plan": {
            "$ref": "#/components/schemas/LicensePlan"
          },
          "status": {
            "$ref": "#/components/schemas/LicenseStatus"
          },
          "seat_limit": {
            "type": "integer",
            "title": "Seat Limit"
          },
          "server_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Limit"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "licensee": {
            "type": "string",
            "title": "Licensee"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan",
          "status",
          "seat_limit",
          "server_limit",
          "issued_at",
          "expires_at",
          "revoked_at",
          "licensee",
          "created_at",
          "updated_at"
        ],
        "title": "LicenseView",
        "description": "Safe response schema for a platform license.\n\nNever exposes ``key_hash`` or the license key itself; the plaintext key is\nreturned only once at issue time via ``LicenseIssueResponse``."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}